Separate Tardis adapter
